improved vagrant installation instructions - and included virtualbox
I noticed the installation instructions on the website and this repository do not mention Virtualbox at all and make it sound like CoreOS runs *on* Vagrant. This confused me at first.

With this rewrite I hope to make it more clear how it actually works. I would also hope to see some to this effect on the http://coreos.com/docs/vagrant/ page. 

http://docs.vagrantup.com/v2/virtualbox/index.html