feat(Vagrantfile) Use box versions, remove monkey patch.
As of 308 we generate json files that include version metadata for the
Vagrant boxes. Also the monkey patching has been moved into the box so
it can be removed from here.

Note: The version constraint ensures new instances will download 308 or
later but this is not enforced on existing instances. Since the monkey
patching is required just to get previous versions of CoreOS to boot
users should not update to this revision if old instances are in use.