Streamlined setup for Windows 7
This is intended to give minimal instructions that are covered in detail in the CoreOS-Vagrant-for-Windows-7-Documentation.docx file provided in the repository.  There are 32 pages of detail in that document.  But we all have minds the size of a planet and egos the size of a universe. Nothing you install cannot be uninstalled and reinstalled cleanly.  This statement is made by experience.