etcd: Add static cluster configuration option
Currently the only way to configure an etcd cluster is by using the
discovery service. This may not always be an option especially for
offline scenarios. An option 'enable_discovery_service' is added to
allow for the enabling/disabling of using the discovery service versus
using a static cluster configuration. This option is defaulted to 'true'
to maintain the use of the discovery service as the default. When the
discovery service option is disabled the etcd cluster will be statically
configured and will be comprised of each instance.

Additionally static configuration requires that each etcd member be
named explicitly. This means that each node requires a 'user-data' file
that has the etcd member name set. When the discovery service option is
disabled a 'user-data' file for each instance will be created. The name
of each of these files will be 'user-data' appended with the instance
id (ex. user-data-01) and will have the etcd member name set to the
instance name (ex. core-01).