etcd: Fix etcd v1 static configuration
Incorrectly assumed that etcd v1 and v2 static configuration was the
same. In v1 a peers list should be provided that is just a list of the
peers ip and port. In addition to a peers list an initial leader must be
selected by not having a peers list. This was accomplished by deleting
the peers list for the last instance of the cluster.