Change VirtualBox NIC
We've seen some instances of the kernel rebooting due
to a kenel paging request in the network RX code, coming
from the virtio_net module.  Switch the second nic, which
is the private network, to be an emulation of an Intel NIC,
avoiding the virtio_net code.

Only the second NIC is changed, the first NIC is left as
a virtio_net NIC.  That NIC is only used for host<->VM
communications, and changing it to an Intel NIC messes up
the NIC configuration, and leaving it as a virtio_net driver
keeps things working.  (The NIC order was getting changed,
causing CoreOS to configure the wrong interface for the
private network.)